+++ web-react-ts/src/components/ChurchGraph/ChurchGraph.tsx
@@ -160,7 +160,12 @@ const ChurchGraph = (props: ChurchGraphProps) => {
                 yAxisId="left"
                 fill="url(#colorPrimary)"
                 onClick={(data) => {
-                  if (data.id === data.week) return
+                  if (
+                    data.category === 'bussingAggregate' ||
+                    data.category === 'serviceAggregate'
+                  ) {
+                    return
+                  }
 
                   if (data.id && bussing) {
                     clickCard({ ...data, __typename: 'BussingRecord' })